Fixed buildButtonArea() bug in window code. Add unit display window (though buggy for now).
git-svn-id: https://spqr.svn.sourceforge.net/svnroot/spqr@137 2be61f21-b1ab-43f6-960c-aab46cf39629

+++ scripts/spqr_window.py
@@ -248,7 +248,7 @@ class CWindow(object):
 			# now we add the buttons
 			xpos = width - ((2 * SPQR.SPACER) + SGFX.gui.iWidth("button"))
 			ypos = (self.rect.h - (extend_height + SPQR.WINSZ_TOP))
-			ypos += (extend_height-SGFX.gui.iHeight("button")/2)
+			ypos += int((extend_height - SGFX.gui.iHeight("button")) / 2)
 			while len(button_list) > 0:	
 				# get the next button
 				button = button_list.pop(0)
